create_class_calendar

+++ src/main/java/com/example/fams/entities/FamsClass.java
@@ -19,7 +19,7 @@ public class FamsClass extends BaseEntity {
 
     private String code;
 
-    private Long duration;
+    private Double duration;
 
     private ClassStatus classStatus;
